RPM: Build static lib package only if --enable_static was enabled.
In the dark ages it was standard to have a package of static libraries.
This has changed and current Fedora policy for example is to not build
or use static libraries anymore if possible.

This package however aims to support a very wide range of distributions
and configurations.  So generate the rpm spec file to not build a static
library package if --disable_static or equivalent was used at configure
time.

The default is to build static libraries.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>